PB-269 : moving feature class out of the features.api.js file
As we will then add a new one, or heavily modify the SelectableFeature class, to implement our generic KML/GPX feature identification
