+++ src/modules/drawing/components/DrawingMarkerInteraction.vue
@@ -3,7 +3,7 @@ import Feature from 'ol/Feature'
 import { computed } from 'vue'
 import { useStore } from 'vuex'
 
-import { EditableFeatureTypes } from '@/api/features.api'
+import { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 import useDrawingModeInteraction from '@/modules/drawing/components/useDrawingModeInteraction.composable'
 
 const emits = defineEmits({
