+++ src/modules/drawing/components/DrawingToolboxButton.vue
@@ -4,7 +4,7 @@ import { computed, toRefs } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useStore } from 'vuex'
 
-import { EditableFeatureTypes } from '@/api/features.api'
+import { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 
 const i18n = useI18n()
 const store = useStore()
