+++ src/modules/infobox/components/styling/SelectedFeatureProfile.vue
@@ -20,7 +20,7 @@
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { mapState } from 'vuex'
 
-import { EditableFeature, EditableFeatureTypes } from '@/api/features.api'
+import EditableFeature, { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 import { geometryInfo } from '@/utils/utils'
 
 export default {
