+++ src/store/modules/drawing.store.js
@@ -1,4 +1,4 @@
-import { EditableFeatureTypes } from '@/api/features.api'
+import { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 import { loadAllIconSetsFromBackend } from '@/api/icon.api'
 
 /**
