+++ src/store/plugins/click-on-map-management.plugin.js
@@ -1,4 +1,4 @@
-import { identify } from '@/api/features.api'
+import { identify } from '@/api/features/features.api.js'
 import LayerTypes from '@/api/layers/LayerTypes.enum'
 import { ClickType } from '@/store/modules/map.store'
 import log from '@/utils/logging'
