+++ src/utils/identifyOnVectorLayer.js
@@ -5,7 +5,8 @@ import { lineString, point, polygon } from '@turf/helpers'
 import pointToLineDistance from '@turf/point-to-line-distance'
 import proj4 from 'proj4'
 
-import { EditableFeature, LayerFeature } from '@/api/features.api'
+import EditableFeature from '@/api/features/EditableFeature.class'
+import LayerFeature from '@/api/features/LayerFeature.class'
 import { WGS84 } from '@/utils/coordinates/coordinateSystems'
 import reprojectGeoJsonData from '@/utils/geoJsonUtils'
 import log from '@/utils/logging'
