+++ tests/cypress/support/drawing.js
@@ -1,6 +1,6 @@
 import pako from 'pako'
 
-import { EditableFeatureTypes } from '@/api/features.api'
+import { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 import { BREAKPOINT_PHONE_WIDTH } from '@/config'
 import { randomIntBetween } from '@/utils/numberUtils'
